safari-600.7-branch/LayoutTests/platform/ios-sim/fast/forms/date/input-date-commit-valid-only-expected.txt

Type=date field should not accept invalid date strings though a user can type such strings

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "2012-04-13"
FAIL document.getSelection().toString() == "+++" should be false. Was true.
PASS input.value is ""
PASS successfullyParsed is true

TEST COMPLETE
